Fix upgrades-test bug, change model
